Move CentralUserCommands to commands folder
