+++ lib/tasks/central_updates_subscriber.rake
@@ -1,5 +1,3 @@
-require './lib/carto/subscribers/central_user_commands'
-
 def process_exists?(pid)
   Process.getpgid(pid)
   true
@@ -31,7 +29,7 @@ namespace :message_broker do
       subscription_name = Carto::Common::MessageBroker::Config.instance.central_subscription_name
       subscription = message_broker.get_subscription(subscription_name)
       notifications_topic = message_broker.get_topic(:cartodb_central)
-      central_user_commands = Carto::Subscribers::CentralUserCommands.new(
+      central_user_commands = CentralUserCommands.new(
         notifications_topic: notifications_topic,
         logger: logger
       )
